Safari-609.1.11.1/LayoutTests/media/modern-media-controls/layout-node/remove-expected.txt

Testing the LayoutNode.remove() method.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


node.children = [a, b, c]
PASS node.children.length === 3 is true
PASS node.children[0] === a is true
PASS node.children[1] === b is true
PASS node.children[2] === c is true

b.remove()
PASS node.children.length === 2 is true
PASS node.children[0] === a is true
PASS node.children[1] === c is true
PASS retVal === b is true

a.remove()
PASS node.children.length === 1 is true
PASS node.children[0] === c is true

Layout was performed
PASS node.element.childElementCount === 1 is true
PASS node.element.firstElementChild === c.element is true
PASS successfullyParsed is true

TEST COMPLETE
